JPA 자동 생성 시 DB 컬럼도 BIGINT로 생성됨
Merge pull request #63 from Thuy111/feature/entity

+++ backend/src/main/java/com/bob/smash/entity/IntroductionImage.java
@@ -37,7 +37,7 @@ public class IntroductionImage {
     private String type;
 
     @Column(nullable = false)
-    private Integer size;
+    private Long size;
 
     public void changeIsMain(Byte isMain) {this.isMain = isMain;}
     public void changeOrderIndex(Integer orderIndex) {this.orderIndex = orderIndex;}
@@ -45,5 +45,5 @@ public class IntroductionImage {
     public void changeOName(String oName) {this.oName = oName;}
     public void changePath(String path) {this.path = path;}
     public void changeType(String type) {this.type = type;}
-    public void changeSize(Integer size) {this.size = size;}
+    public void changeSize(Long size) {this.size = size;}
 }
